// src/app/design-system/projects/spr-components/src/lib/components/period-selector/enums/period-selector-variants.enum.ts
export enum PeriodVariants {
  Today = 'today',
  Yesterday = 'yesterday',
  Week = 'week',
  LastSevenDays = 'last_7_days',
  LastWeek = 'last_week',
  ThisMonth = 'this_month',
  LastMonth = 'last_month',
  LastThirtyDays = 'last_30_days',
  LastThreeMonth = 'last_three_month',
  LastYear = 'last_year',
  Lifetime = 'lifetime',
  All = 'all',
  Custom = 'custom',
  NotSelected = 'notSelected',
}